Chat dropdown (#7)
change chat header from previous and next buttons to a dropdown
(although the next button is still there).

also moves the collapse button to the chat header, instead of floating
over the chat iframe, which caused a "Chatting is disabled for channel
owner/mods because the Twitch Chat window is obscured by another
element" warning that temporarily disabled chatting. this change also
fixes that bug.