Settings page (#9)
Add settings page.

The intent was to add a full screen settings page that could be
navigated to without losing the state of which streams were being
watched. This created the need for an app-wide context. Now, routes that
are do not fall into the catch-all (streams routes) can be created to
reuse the header, sidebar, and anything other app-wide state needed.

- added settings page
- added app context
- added updatePath state (true for catch-all routes only) 
- normalized icon sizing (100%)
- removed hardcoded followed games
- use local storage to save followed games
- updated docs